Added JavaFX and JDBC library
JavaFX and JDBC should be automatically added to the library for the project on every device the project is is pulled to. Check that this is true by going to:
"File/Project Structure -> Libraries -> lib"
It should display the path to the lib folder in the JavaFX + JDBC folder. If not, contact me (Thea) and we will fix it : )

Additionally instantiated main recourses folder, this is where all external files (SQLite, pictures, etc) should go